refactor: rename component-provider.tsx to arte-odyssey-provider.tsx
Update file name to match the component name for consistency.
