+++ packages/arte-odyssey/src/components/providers/index.ts
@@ -1,2 +1,2 @@
-export * from './component-provider';
+export * from './arte-odyssey-provider';
 export * from './portal-root';
